Rename plugin
make the move more dynamic

Format
