+++ cmd/plugin/addon/main.go
@@ -15,13 +15,13 @@ import (
 	"github.com/vmware-tanzu-private/core/pkg/v1/cli"
 	"github.com/vmware-tanzu-private/core/pkg/v1/cli/command/plugin"
 
+	addon "github.com/vmware-tanzu/tce/pkg/addon"
 	cfg "github.com/vmware-tanzu/tce/pkg/common/config"
-	extension "github.com/vmware-tanzu/tce/pkg/extension"
 )
 
 var descriptor = cli.PluginDescriptor{
-	Name:        "extension",
-	Description: "Extension management",
+	Name:        "package",
+	Description: "Package management",
 	Version:     cli.BuildVersion,
 	BuildSHA:    "",
 	Group:       cli.ManageCmdGroup,
@@ -78,13 +78,13 @@ func main() {
 	// p.Cmd.PersistentFlags().StringVar(&logFile, "log_file", "", "Log file path")
 
 	p.AddCommands(
-		extension.ListCmd,
-		extension.GetCmd,
-		extension.ReleaseCmd,
-		extension.InstallCmd,
-		extension.DeleteCmd,
-		extension.TokenCmd,
-		extension.ResetCmd,
+		addon.ListCmd,
+		addon.GetCmd,
+		addon.ReleaseCmd,
+		addon.InstallCmd,
+		addon.DeleteCmd,
+		addon.TokenCmd,
+		addon.ResetCmd,
 	)
 	if err := p.Execute(); err != nil {
 		os.Exit(1)
