+++ pkg/addon/delete.go
@@ -1,7 +1,7 @@
 // Copyright 2020 VMware Tanzu Community Edition contributors. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package extension
+package addon
 
 import (
 	"fmt"
@@ -11,15 +11,15 @@ import (
 	klog "k8s.io/klog/v2"
 )
 
-// InstallCmd represents the install command
-var InstallCmd = &cobra.Command{
-	Use:   "install <extension name>",
-	Short: "Install extension",
+// DeleteCmd represents the delete command
+var DeleteCmd = &cobra.Command{
+	Use:   "delete <extension name>",
+	Short: "Delete extension",
 	PreRunE: func(cmd *cobra.Command, args []string) (err error) {
 		mgr, err = NewManager()
 		return err
 	},
-	RunE: install,
+	RunE: delete,
 	PostRunE: func(cmd *cobra.Command, args []string) (err error) {
 		return nil
 	},
@@ -27,24 +27,27 @@ var InstallCmd = &cobra.Command{
 
 func init() {
 	// common between secret and user-defined
-	InstallCmd.Flags().StringVarP(&inputAppCrd.Namespace, "namespace", "n", "tanzu-extensions", "Namespace to deploy too")
+	DeleteCmd.Flags().StringVarP(&inputAppCrd.Namespace, "namespace", "n", "tanzu-extensions", "Namespace to deploy too")
 
 	// secret
-	InstallCmd.Flags().StringVarP(&inputAppCrd.ClusterName, "cluster", "c", "", "Cluster name which corresponds to a secret")
+	DeleteCmd.Flags().StringVarP(&inputAppCrd.ClusterName, "cluster", "c", "", "Cluster name which corresponds to a secret")
 
 	// user defined
-	InstallCmd.Flags().StringVarP(&inputAppCrd.URL, "url", "u", "", "URL to image")
-	InstallCmd.Flags().StringToStringVarP(&inputAppCrd.Paths, "paths", "p", nil, "User defined paths for kapp template")
+	DeleteCmd.Flags().StringVarP(&inputAppCrd.URL, "url", "u", "", "URL to image")
+	DeleteCmd.Flags().StringToStringVarP(&inputAppCrd.Paths, "paths", "p", nil, "User defined paths for kapp template")
+
+	// delete force
+	DeleteCmd.Flags().BoolVarP(&inputAppCrd.Force, "force", "f", false, "Force delete")
+	DeleteCmd.Flags().BoolVarP(&inputAppCrd.Teardown, "teardown", "t", false, "Delete associated ServiceAccount and RoleBinding")
 }
 
-func install(cmd *cobra.Command, args []string) error {
+func delete(cmd *cobra.Command, args []string) error {
 
 	if len(args) == 0 {
 		fmt.Printf("Please provide extension name\n")
 		return ErrMissingExtensionName
 	}
 	inputAppCrd.Name = args[0]
-	klog.V(2).Infof("install(extension) = %s", inputAppCrd.Name)
 	if inputAppCrd.Name == "" {
 		fmt.Printf("Please provide extension name\n")
 		return ErrMissingExtensionName
@@ -56,7 +59,7 @@ func install(cmd *cobra.Command, args []string) error {
 
 	err := mgr.gh.StageFiles(extensionWorkingDir, inputAppCrd.Name)
 	if err != nil {
-		fmt.Printf("StageFiles failed. Err: %v\n", err)
+		fmt.Printf("DownloadExtension failed. Err: %v\n", err)
 		return err
 	}
 
@@ -77,13 +80,13 @@ func install(cmd *cobra.Command, args []string) error {
 		}
 	*/
 
-	err = mgr.kapp.InstallFromFile(inputAppCrd)
+	err = mgr.kapp.DeleteFromFile(inputAppCrd)
 	if err != nil {
-		fmt.Printf("InstallFromFile failed. Err: %v\n", err)
+		fmt.Printf("Delete failed. Err: %v\n", err)
 		return err
 	}
 
-	fmt.Printf("%s install extension succeeded\n", inputAppCrd.Name)
+	fmt.Printf("%s delete extension succeeded\n", inputAppCrd.Name)
 
 	return nil
 }
