+++ pkg/addon/release.go
@@ -1,7 +1,7 @@
 // Copyright 2020 VMware Tanzu Community Edition contributors. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package extension
+package addon
 
 import (
 	"fmt"
